Split scenario-workspace out of its router (531 -> 54 lines)
scenario-interact had both the configure and play phases inlined, so the
component whose job is choosing between phases was 531 lines. dream-interact
does the same job in 57, because its markup and its logic live in
dream-workspace. That is the frame:

  <x-gallery v-if="!selected" />     <x-workspace v-else />

Nothing about how a Scenario is configured or played changes here -- the
template and script moved across intact. Per Silas, the interact tier "is
where we are actually hitting what we do with them uniquely", so the shared
thing is the frame, not the contents.

Two details worth stating:

  - The status banner STAYS in the router. It reports on browse, configure
    and play alike, so it belongs to the frame, the same way kr-manager owns
    its one banner.
  - No class is passed to <scenario-workspace>. It is multi-root (configure
    and play are siblings), so an attribute would have nowhere to land; both
    sections already carry their own `min-h-0 flex-1`.

Rule 5 records the drop: scenario-interact 531 -> 54 (-477).

vue-tsc clean, eslint clean, route-gallery/reachability/layout/wonderlab
--strict all green.
