+++ utils/scripts/route-gallery-baseline.json
@@ -28,6 +28,6 @@
     "dream-interact": 57,
     "facet-interact": 204,
     "reward-interact": 1097,
-    "scenario-interact": 531
+    "scenario-interact": 54
   }
 }
